Refactored plugin common code
 - added plugins root include path
 - added missing pokey plugin main.cpp file
 - split PluginStateManager base class into private shared
   header file
 - added specific derivations of base class in each of the two
   current plugin impelmentations
