+++ src/libs/plugins/common/simhubdeviceplugin.h
