+++ src/libs/plugins/simsource/simsource_main.h
@@ -1,0 +1,11 @@
+#ifndef __SIMSOURCE_MAIN_H
+#define __SIMSOURCE_MAIN_H
+
+#include "common/private/pluginstatemanager.h"
+
+//! barest specialisation of the internal plugin management support base class
+class SimSourcePluginStateManager : public PluginStateManager
+{
+};
+
+#endif
